tests/wpt/web-platform-tests/credential-management/fedcm-multi-idp/get-before-onload-and-during-dom-content-loaded.https.html

<!DOCTYPE html>
<title>Federated Credential Management API multi IDP get before onload and during DOMContentLoaded test.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>

<script type="module">
import {set_fedcm_cookie, set_alt_fedcm_cookie, default_request_options, default_alt_request_options} from '../support/fedcm-helper.sub.js';

let cookies_promise = Promise.all([set_fedcm_cookie(), set_alt_fedcm_cookie()]);

promise_test(async t => {
  let has_dom_content_loaded = false;
  const dom_content_loaded = new Promise(resolve => {
      document.addEventListener('DOMContentLoaded', async () => {
          const second_cred = navigator.credentials.get(default_alt_request_options());
          await promise_rejects_dom(t, 'NetworkError', second_cred);
          has_dom_content_loaded = true;
          resolve();
      });
  });
  assert_false(has_dom_content_loaded);
  const first_cred = navigator.credentials.get(default_request_options());
  await Promise.all([cookies_promise, dom_content_loaded]);
  assert_true(has_dom_content_loaded);
  const first = await first_cred;
  assert_equals(first.token, "token");
}, "A `get` call before onload and a `get` call during DOMContentLoaded event should combine despite being called from different tasks.");

</script>
